+++ crypto_pandas/ccxt/ccxt_pandas_exchange.py
@@ -27,87 +27,46 @@ from crypto_pandas.utils.pandas_utils import (
     preprocess_order_dataframe,
 )
 
-ccxt_processor = BaseProcessor()
-
 
 @dataclass
 class CCXTPandasExchange:
     """
-    A wrapper class for a CCXT exchange with extended functionalities using pandas.
-
-    This class extends the CCXT `Exchange` class and provides additional methods
-    for working with cryptocurrency exchanges. It integrates with pandas DataFrames
-    for processing and presenting exchange data as well as preprocessing and
-    validation mechanisms for orders.
+    CCXTPandasExchange is a wrapper for the CCXT library that integrates with Pandas
+    to provide streamlined data processing for cryptocurrency exchanges. It enables users
+    to seamlessly create orders, fetch market data, and process exchange responses as
+    Pandas DataFrames.
 
     Attributes:
-        exchange (Exchange): Instance of a CCXT exchange, defaulting to Binance.
-        max_order_notional (float): Upper limit for the total notional value of an order.
-            Default is 10,000.
-        max_number_of_orders (int): Maximum number of allowed orders in a batch.
-            Default is 5.
-        markets_cache_time (int): Time to cache market data in seconds. Default is 86400
-            seconds (1 day).
-        order_amount_rounding (Literal["floor", "ceil", "round"]): Strategy for rounding
-            order amounts. Default is "round".
-        order_price_rounding (Literal["aggressive", "defensive", "round"]): Strategy for
-            rounding order prices. Default is "round".
+        exchange (ccxt.Exchange): An instance of the CCXT exchange client.
+        exchange_name (str | None): The name of the exchange to interact with.
+        account_name (str | None): The account name, if required for tracking.
+        max_order_notional (float): Maximum notional value for any single order.
+        max_number_of_orders (int): Maximum number of bulk orders allowed.
+        markets_cache_time (int): Cache duration (in seconds) for markets data.
+        order_amount_rounding (Literal["floor", "ceil", "round"]): Strategy for rounding order amounts.
+        order_price_rounding (Literal["aggressive", "defensive", "round"]): Strategy for rounding order prices.
+        _ccxt_processor (BaseProcessor): A helper class to process CCXT responses and provide consistent output.
 
-        Methods:
-        load_markets: Load market data and cache it.
-        fetch_balance: Fetch account balance as a DataFrame.
-        fetch_trading_fee: Fetch trading fee as a DataFrame.
-        fetch_trading_fees: Fetch trading fees as a DataFrame.
-        fetch_positions_risk: Fetch risk information for open positions.
-        fetch_position: Fetch details of a specific position.
-        fetch_positions: Fetch all positions.
-        fetch_transfers: Fetch transfer history.
-        fetch_ledger: Fetch the account ledger.
-        fetch_withdrawals: Fetch withdrawal history.
-        fetch_currencies: Fetch supported currencies.
-        fetch_ticker: Fetch ticker information for a specific symbol.
-        fetch_tickers: Fetch all tickers as a DataFrame.
-        fetch_order_book: Fetch order book for a given symbol.
-        fetch_ohlcv: Fetch OHLCV (candlestick) data.
-        fetch_funding_history: Fetch historical funding data.
-        fetch_funding_rate_history: Fetch historical funding rates.
-        fetch_open_interest: Fetch open interest for a given symbol.
-        fetch_open_interest_history: Fetch historical open interest.
-        fetch_status: Fetch the exchange status.
-        fetch_trades: Fetch recent trades for a given symbol.
-        fetch_my_trades: Fetch user's trade history.
-        fetch_leverages: Fetch leverage settings.
-        fetch_liquidations: Fetch liquidation events.
-        fetch_greeks: Fetch greek values (options metrics).
-        fetch_long_short_ratio_history: Fetch historical long-short ratios.
-        fetch_margin_adjustment_history: Fetch historical margin adjustment data.
-        fetch_my_liquidations: Fetch user's liquidation history.
-        fetch_option: Fetch option details.
-        fetch_funding_rates: Fetch funding rates.
-        fetch_convert_trade_history: Fetch historical convert trades.
-        fetch_bids_asks: Fetch the latest bid/ask prices for multiple symbols.
-        fetch_orders: Fetch order history.
-        fetch_open_orders: Fetch open orders.
-        fetch_closed_orders: Fetch closed orders.
-        fetch_canceled_and_closed_orders: Fetch canceled and closed orders.
-        fetch_order: Fetch details of a specific order.
-        create_order: Create a new order after preprocessing.
-        cancel_order: Cancel a given order by ID.
-        edit_order: Edit an existing order with new parameters.
-        create_orders: Batch-create multiple orders.
-        cancel_orders: Cancel a batch of orders using their IDs.
-        cancel_all_orders: Cancel all orders for a specific symbol or account-wide.
-        cancel_orders_for_symbols: Cancel specific orders across symbols.
-        edit_orders: Batch-edit multiple orders.
-
+    Methods:
+        __getattr__(method_name: str): Overridden to enable dynamic method resolution for CCXT methods,
+                                       with transformations applied to handle inputs and outputs as Pandas DataFrames.
+        load_cached_markets(params: dict = {}): Loads and caches market data from the exchange.
     """
 
     exchange: ccxt.Exchange = field(default_factory=ccxt.binance)
+    exchange_name: str | None = None
+    account_name: str | None = None
     max_order_notional: float = 10_000
     max_number_of_orders: int = 5
     markets_cache_time: int = 3600
     order_amount_rounding: Literal["floor", "ceil", "round"] = "round"
     order_price_rounding: Literal["aggressive", "defensive", "round"] = "round"
+    _ccxt_processor: BaseProcessor = field(default_factory=BaseProcessor)
+
+    def __post_init__(self):
+        self._ccxt_processor = BaseProcessor(
+            exchange_name=self.exchange_name, account_name=self.account_name
+        )
 
     def __getattr__(self, method_name: str) -> Callable:
         original_method = getattr(self.exchange, method_name)
@@ -142,30 +101,30 @@ class CCXTPandasExchange:
                     price_strategy=self.order_price_rounding,
                     amount_strategy=self.order_amount_rounding,
                 )
-                kwargs["orders"] = ccxt_processor.orders_to_dict(
+                kwargs["orders"] = self._ccxt_processor.orders_to_dict(
                     orders=kwargs["orders"]
                 )
             elif method_name in symbol_order_methods:
                 kwargs["orders"] = kwargs["orders"][["id", "symbol"]].to_dict("records")
             result = original_method(*args, **kwargs)
             if method_name in standard_dataframe_methods:
-                result = ccxt_processor.response_to_dataframe(data=result)
+                result = self._ccxt_processor.response_to_dataframe(data=result)
             elif method_name in markets_dataframe_methods:
-                result = ccxt_processor.markets_to_dataframe(data=result)
+                result = self._ccxt_processor.markets_to_dataframe(data=result)
             elif method_name in balance_dataframe_methods:
-                result = ccxt_processor.balance_to_dataframe(data=result)
+                result = self._ccxt_processor.balance_to_dataframe(data=result)
             elif method_name in ohlcv_dataframe_methods:
-                result = ccxt_processor.ohlcv_to_dataframe(
+                result = self._ccxt_processor.ohlcv_to_dataframe(
                     data=result, symbol=kwargs.get("symbol")
                 )
             elif method_name in orderbook_dataframe_methods:
-                result = ccxt_processor.order_book_to_dataframe(data=result)
+                result = self._ccxt_processor.order_book_to_dataframe(data=result)
             elif method_name in orders_dataframe_methods:
-                result = ccxt_processor.orders_to_dataframe(data=result)
+                result = self._ccxt_processor.orders_to_dataframe(data=result)
             elif method_name in ohlcv_symbols_dataframe_methods:
-                result = ccxt_processor.ohlcv_symbols_to_dataframe(data=result)
+                result = self._ccxt_processor.ohlcv_symbols_to_dataframe(data=result)
             elif method_name in dict_methods:
-                result = ccxt_processor.preprocess_dict(data=result)
+                result = self._ccxt_processor.preprocess_dict(data=result)
             return result
 
         return wrapped